+++ src/dummywallet.cpp
@@ -32,3 +32,24 @@ void DummyWalletInit::AddWalletOptions() const {
 const WalletInitInterface &g_wallet_init_interface = DummyWalletInit();
 
 void StopGenerateThread() {}
+fs::path GetWalletDir() {
+    throw std::logic_error("Wallet function called in non-wallet build.");
+}
+
+std::vector<fs::path> ListWalletDir() {
+    throw std::logic_error("Wallet function called in non-wallet build.");
+}
+
+std::vector<std::shared_ptr<CWallet>> GetWallets() {
+    throw std::logic_error("Wallet function called in non-wallet build.");
+}
+
+namespace interfaces {
+
+class Wallet;
+
+std::unique_ptr<Wallet> MakeWallet(const std::shared_ptr<CWallet> &wallet) {
+    throw std::logic_error("Wallet function called in non-wallet build.");
+}
+
+} // namespace interfaces