+++ doc/release-notes.md
@@ -3,3 +3,6 @@ This release includes the following features and fixes since 1.2.1 release:
  - Fix bug in rpc method verifymessage
  - increase min change (+ min final change)
  - Change some CMake build defaults
+ - `listwalletdir` returns a list of wallets in the wallet directory which is
+   configured with `-walletdir` parameter.
+